+++ src/main/java/roomescape/domain/ReservationTimeRepository.java
@@ -1,6 +1,4 @@
-package roomescape.repository;
-
-import roomescape.model.ReservationTime;
+package roomescape.domain;
 
 import java.util.List;
