+++ src/main/java/roomescape/dto/ReservationTimeResponse.java
@@ -1,6 +1,6 @@
 package roomescape.dto;
 
-import roomescape.model.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 import java.time.LocalTime;
